requirements.txt
langid==1.1.6
PyGithub==2.8.1
PyYAML==6.0.3